+++ docs/releases.rst
@@ -16,8 +16,8 @@ Graphite-API releases
   database-backed finders such as Cyanite because it allows fetching all time
   series at once instead of sequentially.
 
-* Add ``multiplySeriesWithWildcards``, ``minimumBelow`` and ``changed``
-  functions.
+* Add ``multiplySeriesWithWildcards``, ``minimumBelow``, ``changed``,
+  ``timeSlice`` and ``removeEmptySeries`` functions.
 
 * Add optional ``step`` argument to ``time``, ``sin`` and ``randomWalk``
   functions.